src/pyproject.toml
[project]
name = "hybrid-microservice-adk-agent-observability"
version = "0.1.0"
description = "An example of joining spans together under a common trace ID when dealing with a hybrid 'microservice -> ADK agent -> microservice'-type topology."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.116.1",
    "google-adk==1.11.0",
    "loguru==0.7.2",
    "pydantic-settings==2.5.2",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-gcp-monitoring==1.9.0a0",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-instrumentation==0.57b0",
    "opentelemetry-instrumentation-fastapi==0.57b0",
    "opentelemetry-instrumentation-httpx==0.57b0",
    "opentelemetry-propagator-gcp==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.36.0",
    "uv==0.8.11",
    "uvicorn==0.35.0",
]
